more tests, improvements to minutes, improvements to DateResolutionExt
